conda_requirements.txt
numpy
pytorch >= 2.0
torchtriton
matplotlib
numba
cupy
ase
h5py
tqdm
python-graphviz
lightning
opt_einsum